api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.26.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
Pillow>=8.3.1
paddlepaddle>=2.2.0
paddleocr>=2.4.0
torch>=1.9.0
transformers>=4.9.0
